+++ db/migrate/20200111155410_devise_create_admin_users.rb
@@ -4,6 +4,8 @@ class DeviseCreateAdminUsers < ActiveRecord::Migration[6.0]
   def change
     create_table :admin_users do |t|
       t.references :organization, null: true, foreign_key: true
+      t.string :firstname, limit: 30
+      t.string :lastname, limit: 30
 
       ## Database authenticatable
       t.string :email,              null: false, default: ""
@@ -17,11 +19,11 @@ class DeviseCreateAdminUsers < ActiveRecord::Migration[6.0]
       t.datetime :remember_created_at
 
       ## Trackable
-      # t.integer  :sign_in_count, default: 0, null: false
-      # t.datetime :current_sign_in_at
-      # t.datetime :last_sign_in_at
-      # t.inet     :current_sign_in_ip
-      # t.inet     :last_sign_in_ip
+      t.integer  :sign_in_count, default: 0, null: false
+      t.datetime :current_sign_in_at
+      t.datetime :last_sign_in_at
+      t.inet     :current_sign_in_ip
+      t.inet     :last_sign_in_ip
 
       ## Confirmable
       # t.string   :confirmation_token