+++ db/schema.rb
@@ -31,11 +31,18 @@ ActiveRecord::Schema.define(version: 2020_03_04_182631) do
 
   create_table "admin_users", force: :cascade do |t|
     t.bigint "organization_id"
+    t.string "firstname", limit: 30
+    t.string "lastname", limit: 30
     t.string "email", default: "", null: false
     t.string "encrypted_password", default: "", null: false
     t.string "reset_password_token"
     t.datetime "reset_password_sent_at"
     t.datetime "remember_created_at"
+    t.integer "sign_in_count", default: 0, null: false
+    t.datetime "current_sign_in_at"
+    t.datetime "last_sign_in_at"
+    t.inet "current_sign_in_ip"
+    t.inet "last_sign_in_ip"
     t.integer "roles_mask"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false